requirements.txt
datasets==4.1.1
faiss==1.9.0
fastapi==0.118.0
neo4j==5.28.2
networkx==3.5
numpy==2.3.3
openai==2.0.1
pandas==2.3.3
pydantic==2.11.9
python-dotenv==1.1.1
scikit_learn==1.7.2
sentence_transformers==5.1.1
torch==2.6.0+cu124
transformers==4.56.2
trl==0.23.0
uvicorn==0.37.0
weaviate_client==4.17.0
